SWAGGER 처리

+++ src/main/java/com/org/qualitycore/work/model/service/WorkService.java
@@ -1,17 +1,20 @@
 package com.org.qualitycore.work.model.service;
 
+import com.org.qualitycore.exception.ResourceNotFoundException;
 import com.org.qualitycore.work.model.dto.WorkDTO;
 import com.org.qualitycore.work.model.entity.Work;
 import com.org.qualitycore.work.model.repository.WorkRepository;
+import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
-
+import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
+@Schema(description = "작업지시서 관련 Service")
 public class WorkService {
 
     private final WorkRepository workRepository;
@@ -33,14 +36,16 @@ public class WorkService {
         return modelMapper.map(work, WorkDTO.class);
     }
 
+    @Transactional
     public void workOrderCreate(WorkDTO work) {
 
         workRepository.save(modelMapper.map(work, Work.class));
     }
 
+    @Transactional
     public void workOrderUpdate(WorkDTO work) {
 
-        Work workUpdate = workRepository.findById(work.getWorkOrderId()).orElseThrow(IllegalArgumentException::new);
+        Work workUpdate = workRepository.findById(work.getWorkOrderId()).orElseThrow(() -> new IllegalArgumentException("작업지시서를 찾을 수 없습니다."));
 
         Work workUpdates = workUpdate.toBuilder().
                            workTeam(work.getWorkTeam()).
@@ -55,10 +60,15 @@ public class WorkService {
         workRepository.save(workUpdates);
     }
 
+    @Transactional
     public void workOrderDelete(int workId) {
+        // 작업지시서가 존재하는지 확인
+        if (!workRepository.existsById(workId)) {
+            throw new ResourceNotFoundException("작업지시서를 찾을 수 없습니다.");
+        }
 
+        // 작업지시서가 존재하면 삭제
         workRepository.deleteById(workId);
-
-        modelMapper.map(workId, Work.class);
     }
+
 }
